+++ R/ApplyDissimilaridade.R
@@ -30,7 +30,7 @@
 #' data(Dados.FMI.Quali)
 #' DadosQuali=ApplyDissimilaridade(Dados.FMI.Quali[,6:10],Dados.FMI.Quali[,2])
 #' Dist=Distancia(DadosQuali,1)
-#' Dendograma(Dist, 3)
+#' Dendo=Dendograma(Dist, 3)
 #' @importFrom grDevices colorRampPalette
 #' @export
 #'