+++ R/Dados.DBC.Misto.R
@@ -1,0 +1,20 @@
+#' Dados de exemplo de um experimento em DBC com dados quantitativos e mistos.
+#'
+#' @description Exemplo de parte de um exeperimento com clones batata-doce em DBC
+#' com dados quantitativos e qualitativos.
+#'
+#' @docType data
+#' @keywords datasets
+#' @name Dados.DBC.Misto
+#' @usage data(Dados.DBC.Misto)
+#' @format Um data frame. A primeira coluna
+#' a identificacao dos tratamentos, a segunda  a
+#' identificacao de blocos e da terceira em diante as variaveis respostas
+#'  (dependentes).
+#'
+#' @references
+#' <https://www.youtube.com/playlist?list=PLvth1ZcREyK72M3lFl7kBaHiVh5W53mlR>
+#' @author Alcinei Mistico Azevedo: \email{alcineimistico@@hotmail.com}
+NULL
+
+